projects/instagram_watermark_remover/requirements.txt
# Google Photos API
google-api-python-client==2.114.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Image Processing
Pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80

# IOPaint (Lama Cleaner) for watermark removal
iopaint==1.3.2
torch
torchvision

# Utilities
requests==2.31.0
tqdm==4.66.1
python-dotenv==1.0.0
